mlflow example fixup
