+++ docker/tune_test/Dockerfile
@@ -16,6 +16,7 @@ RUN pip install --upgrade git+git://github.com/hyperopt/hyperopt.git
 RUN pip install --upgrade sigopt
 RUN pip install --upgrade nevergrad
 RUN pip install --upgrade scikit-optimize
+RUN pip install -U mlflow
 RUN pip install -U pytest-remotedata>=0.3.1
 RUN conda install pytorch-cpu torchvision-cpu -c pytorch
